Reuse test for DataAvailability and IndexAvailability
