+++ src/dbnode/storage/bootstrap/bootstrapper/peers/source_data_test.go
@@ -134,181 +134,6 @@ func TestPeersSourceCan(t *testing.T) {
 	assert.False(t, src.Can(bootstrap.BootstrapParallel))
 }
 
-type sourceAvailableHost struct {
-	host        string
-	shards      []uint32
-	shardStates shard.State
-}
-
-func TestPeersSourceAvailableData(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	replicaMajority := 2
-	blockSize := 2 * time.Hour
-	nsMetadata := testNamespaceMetadata(t)
-	shards := []uint32{0, 1, 2, 3}
-	shardTimeRangesToBootstrap := result.ShardTimeRanges{}
-	blockStart := time.Now().Truncate(blockSize)
-	for _, shard := range shards {
-		shardTimeRangesToBootstrap[shard] = xtime.Ranges{}.AddRange(xtime.Range{
-			Start: blockStart,
-			End:   blockStart.Add(blockSize),
-		})
-	}
-
-	testCases := []struct {
-		title                             string
-		hosts                             []sourceAvailableHost
-		bootstrapReadConsistency          topology.ReadConsistencyLevel
-		expectedAvailableShardsTimeRanges result.ShardTimeRanges
-	}{
-		{
-			title: "Returns empty if only self is available",
-			hosts: []sourceAvailableHost{
-				sourceAvailableHost{
-					host:        "self",
-					shards:      shards,
-					shardStates: shard.Available,
-				},
-			},
-			bootstrapReadConsistency:          topology.ReadConsistencyLevelMajority,
-			expectedAvailableShardsTimeRanges: result.ShardTimeRanges{},
-		},
-		{
-			title: "Returns empty if all other peers initializing",
-			hosts: []sourceAvailableHost{
-				sourceAvailableHost{
-					host:        "self",
-					shards:      shards,
-					shardStates: shard.Available,
-				},
-				sourceAvailableHost{
-					host:        "other1",
-					shards:      shards,
-					shardStates: shard.Initializing,
-				},
-				sourceAvailableHost{
-					host:        "other2",
-					shards:      shards,
-					shardStates: shard.Initializing,
-				},
-			},
-			bootstrapReadConsistency:          topology.ReadConsistencyLevelMajority,
-			expectedAvailableShardsTimeRanges: result.ShardTimeRanges{},
-		},
-		{
-			title: "Returns success if consistency can be met",
-			hosts: []sourceAvailableHost{
-				sourceAvailableHost{
-					host:        "self",
-					shards:      shards,
-					shardStates: shard.Initializing,
-				},
-				sourceAvailableHost{
-					host:        "other1",
-					shards:      shards,
-					shardStates: shard.Available,
-				},
-				sourceAvailableHost{
-					host:        "other2",
-					shards:      shards,
-					shardStates: shard.Available,
-				},
-			},
-			bootstrapReadConsistency:          topology.ReadConsistencyLevelMajority,
-			expectedAvailableShardsTimeRanges: shardTimeRangesToBootstrap,
-		},
-		{
-			title: "Returns empty if consistency can not be met",
-			hosts: []sourceAvailableHost{
-				sourceAvailableHost{
-					host:        "self",
-					shards:      shards,
-					shardStates: shard.Initializing,
-				},
-				sourceAvailableHost{
-					host:        "other1",
-					shards:      shards,
-					shardStates: shard.Available,
-				},
-				sourceAvailableHost{
-					host:        "other2",
-					shards:      shards,
-					shardStates: shard.Available,
-				},
-			},
-			bootstrapReadConsistency:          topology.ReadConsistencyLevelAll,
-			expectedAvailableShardsTimeRanges: result.ShardTimeRanges{},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.title, func(t *testing.T) {
-			hostShardSets := []topology.HostShardSet{}
-
-			for _, host := range tc.hosts {
-				shards := sharding.NewShards(host.shards, host.shardStates)
-				shardSet, err := sharding.NewShardSet(shards, sharding.DefaultHashFn(0))
-				require.NoError(t, err)
-
-				hostShardSet := topology.NewHostShardSet(
-					topology.NewHost(host.host, host.host+"address"), shardSet)
-				hostShardSets = append(hostShardSets, hostShardSet)
-			}
-
-			mockMap := topology.NewMockMap(ctrl)
-			mockMap.EXPECT().HostShardSets().Return(hostShardSets).AnyTimes()
-			mockMap.EXPECT().MajorityReplicas().Return(replicaMajority).AnyTimes()
-
-			mockTopology := topology.NewMockTopology(ctrl)
-			mockTopology.EXPECT().
-				Get().
-				Return(mockMap)
-
-			mockAdminSession := client.NewMockAdminSession(ctrl)
-			mockAdminSession.EXPECT().
-				Topology().
-				Return(mockTopology, nil)
-
-			mockClient := client.NewMockAdminClient(ctrl)
-			mockClient.EXPECT().
-				DefaultAdminSession().
-				Return(mockAdminSession, nil)
-			mockClient.EXPECT().
-				Options().
-				Return(
-					client.NewAdminOptions().
-						SetOrigin(topology.NewHost("self", "selfAddress")),
-				)
-
-			mockRuntimeOpts := m3dbruntime.NewMockOptions(ctrl)
-			mockRuntimeOpts.
-				EXPECT().
-				ClientBootstrapConsistencyLevel().
-				Return(tc.bootstrapReadConsistency).
-				AnyTimes()
-
-			mockRuntimeOptsMgr := m3dbruntime.NewMockOptionsManager(ctrl)
-			mockRuntimeOptsMgr.
-				EXPECT().
-				Get().
-				Return(mockRuntimeOpts).
-				AnyTimes()
-
-			opts := testDefaultOpts.
-				SetAdminClient(mockClient).
-				SetRuntimeOptionsManager(mockRuntimeOptsMgr)
-
-			src, err := newPeersSource(opts)
-			require.NoError(t, err)
-
-			res := src.AvailableData(nsMetadata, shardTimeRangesToBootstrap)
-			require.Equal(t, tc.expectedAvailableShardsTimeRanges, res)
-		})
-	}
-}
-
 func TestPeersSourceEmptyShardTimeRanges(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
